requirements.txt
DateTime==5.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
pandas==2.2.1
psutil==5.9.8
psycopg2-binary==2.9.9
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
sqlparse==0.4.4